Rearrange namespaces to make things clearer on outbox
