+++ src/TbdDevelop.Mediator.Outbox.SqlServer/Extensions/MediatorOutboxConfigurationBuilderExtensions.cs
@@ -1,8 +1,8 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using TbdDevelop.Mediator.Outbox.Contracts;
 using TbdDevelop.Mediator.Outbox.Extensions.Configuration;
+using TbdDevelop.Mediator.Outbox.Infrastructure;
 using TbdDevelop.Mediator.Outbox.SqlServer.Context;
 
 namespace TbdDevelop.Mediator.Outbox.SqlServer.Extensions;
