+++ src/TbdDevelop.Mediator.Outbox.SqlServer/SqlServerOutboxStorage.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json;
 using Microsoft.EntityFrameworkCore;
-using TbdDevelop.Mediator.Outbox.Contracts;
+using TbdDevelop.Mediator.Outbox.Infrastructure;
 using TbdDevelop.Mediator.Outbox.SqlServer.Context;
 using TbdDevelop.Mediator.Outbox.SqlServer.Models;
