+++ src/TbdDevelop.Mediator.Outbox/Extensions/Configuration/MediatorOutboxConfigurationBuilder.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
-using TbdDevelop.Mediator.Outbox.Contracts;
+using TbdDevelop.Mediator.Outbox.Infrastructure;
 using TbdDevelop.Mediator.Outbox.Outbox;
 using TbdDevelop.Mediator.Outbox.Services;
