+++ src/TbdDevelop.Mediator.Outbox/Infrastructure/IOutbox.cs
@@ -1,6 +1,6 @@
 ﻿using Mediator;
 
-namespace TbdDevelop.Mediator.Outbox.Contracts;
+namespace TbdDevelop.Mediator.Outbox.Infrastructure;
 
 public interface IOutbox
 {
