+++ src/TbdDevelop.Mediator.Outbox/Infrastructure/IQueueProcessor.cs
@@ -1,0 +1,8 @@
+﻿using TbdDevelop.Mediator.Outbox.Services;
+
+namespace TbdDevelop.Mediator.Outbox.Infrastructure;
+
+public interface IQueueProcessor
+{
+    Task<QueueStatus> ProcessNextOutboxQueueMessage(CancellationToken cancellationToken);
+}
