+++ src/TbdDevelop.Mediator.Outbox/Outbox/DefaultOutboxProcessingPublisher.cs
@@ -1,5 +1,5 @@
 ﻿using Mediator;
-using TbdDevelop.Mediator.Outbox.Contracts;
+using TbdDevelop.Mediator.Outbox.Infrastructure;
 
 namespace TbdDevelop.Mediator.Outbox.Outbox;
