+++ src/TbdDevelop.Mediator.Outbox/Outbox/InMemoryOutboxStorage.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Concurrent;
 using Mediator;
-using TbdDevelop.Mediator.Outbox.Contracts;
+using TbdDevelop.Mediator.Outbox.Infrastructure;
 
 namespace TbdDevelop.Mediator.Outbox.Outbox;
